+++ app/controllers/linebot_controller.rb
@@ -1,4 +1,6 @@
 class LinebotController < ApplicationController
+    require 'line/bot'
+
     skip_before_action :verify_authenticity_token
 
     def callback
@@ -19,7 +21,19 @@ class LinebotController < ApplicationController
         end
     end
 
-    def push_message
-        Notification.send_notifications
+    def push_message(user_id, message)
+        client.push_message(user_id, {
+            type: 'text',
+            text: message
+        })
+    end
+
+    private
+
+    def client
+        @client ||= Line::Bot::Client.new { |config|
+        config.channel_secret = ENV["LINE_BOT_CHANNEL_SECRET"]
+        config.channel_token = ENV["LINE_BOT_CHANNEL_TOKEN"]
+        }
     end
 end
