+++ app/models/notification.rb
@@ -13,13 +13,35 @@ class Notification < ApplicationRecord
     self.update(next_notification_day: new_notification_day, notification_interval: interval)
   end
 
+  def self.send_notifications
+    notifications = Notification.date_of_notification
+    notifications.each do |notification|
+      user = notification.item&.user
+      Rails.logger.debug("Notification ID: #{notification.id}, User: #{user.inspect}, UID: #{user&.uid}")
+      line_user_id = user&.uid
+      if line_user_id.present?
+        Rails.logger.debug("メッセージ作成開始: Notification ID #{notification.id}")
+        # インスタンスメソッドとして呼び出す
+        message = notification.item.create_notification_message
+        Rails.logger.debug("生成された通知メッセージ: #{message}")
+        begin
+          LinebotController.new.push_message(line_user_id, message)
+        rescue => error
+          Rails.logger.error("LINE通知送信エラー: #{error.message}")
+        end
+      else
+        Rails.logger.error("LINE通知送信エラー: UIDが見つかりません (Notification ID: #{notification.id})")
+      end
+    end
+  end
+
   private
 
   # カスタムバリデーション
   def valid_update_next_notification_day
     if next_notification_day.blank?
       errors.add(:next_notification_day, :blank_field)
-    elsif next_notification_day < Date.today + 1.days
+    elsif next_notification_day < Date.today# + 1.days
       errors.add(:next_notification_day, :past_date)
     elsif next_notification_day > Date.today + 365.days
       errors.add(:next_notification_day, :too_far)
@@ -28,27 +50,5 @@ class Notification < ApplicationRecord
 
   def self.date_of_notification
     where(next_notification_day: Date.today)
-  end
-
-  def create_notification_message
-    message = "#{item.name}の残量が少なくなっています\n"
-    message += " メモ: #{item.memo}\n" if item.memo.present?
-    message += "https://stockmate-a7c103b7b0ba.herokuapp.com/\n"
-    message
-  end
-
-  def self.send_notifications
-    notifications = Notification.date_of_notification
-    notifications.each do |notification|
-    message = notification.create_notification_message
-      begin
-        client.push_message(notification.item.user.line_user_id, {
-          type: 'text',
-          text: message
-        })
-      rescue => error
-        Rails.logger.error("LINE通知送信エラー: #{error.message}")
-      end
-    end
   end
 end
